+++ Nos/Models/CoreData/NosNotification+CoreDataClass.swift
@@ -44,13 +44,14 @@ public class NosNotification: NSManagedObject {
         return fetchRequest
     }
     
-    static func unreadCount(for user: Author, in context: NSManagedObjectContext) throws -> Int {
+    static func unreadCount(in context: NSManagedObjectContext) throws -> Int {
         let fetchRequest = NSFetchRequest<NosNotification>(entityName: String(describing: NosNotification.self))
         fetchRequest.predicate = NSPredicate(format: "isRead != 1")
         return try context.count(for: fetchRequest)
     }
     
-    static func markAllAsRead(for user: Author, in context: NSManagedObjectContext) async throws {
+    // TODO: user is unused; is this a bug?
+    static func markAllAsRead(in context: NSManagedObjectContext) async throws {
         try await context.perform {
             let fetchRequest = NSFetchRequest<NosNotification>(entityName: String(describing: NosNotification.self))
             fetchRequest.predicate = NSPredicate(format: "isRead != 1")